Rename XdrLargeInt -> Int

+++ src/numbers/index.js
@@ -4,4 +4,4 @@ export { Uint128 } from './uint128';
 export { Uint256 } from './uint256';
 export { Int128 } from './int128';
 export { Int256 } from './int256';
-export { XdrLargeInt };
+export { XdrLargeInt as Int };
